Classes from 'utils' package moved to 'core' by respective packages
