+++ sdk/src/main/kotlin/com/processout/sdk/api/model/response/POCustomerAction.kt
@@ -1,6 +1,6 @@
 package com.processout.sdk.api.model.response
 
-import com.processout.sdk.utils.findBy
+import com.processout.sdk.core.util.findBy
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
