+++ sdk/src/main/kotlin/com/processout/sdk/api/model/threeds/PO3DS2Configuration.kt
@@ -1,6 +1,6 @@
 package com.processout.sdk.api.model.threeds
 
-import com.processout.sdk.utils.findBy
+import com.processout.sdk.core.util.findBy
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
