+++ sdk/src/main/kotlin/com/processout/sdk/core/util/EnumExtensions.kt
@@ -1,4 +1,4 @@
-package com.processout.sdk.utils
+package com.processout.sdk.core.util
 
 internal inline infix fun <reified E : Enum<E>, V> ((E) -> V).findBy(value: V): E? {
     return enumValues<E>().firstOrNull { this(it) == value }
